requirements.txt
webdriver-manager==4.0.1
python-dotenv
selenium==4.16.0
pandas==2.1.4